nr: Compute ARFCN values from frequency, and vice-versa

+++ model/nr-ue-net-device.h
@@ -194,7 +194,7 @@ class NrUeNetDevice : public NrNetDevice
     NS_DEPRECATED("Obsolete method")
     void UpdateConfig();
 
-    uint16_t GetArfcn(uint8_t index) const;
+    uint32_t GetArfcn(uint8_t index) const;
 
   protected:
     // inherited from Object
