+++ test/nr-test-rrc.cc
@@ -477,8 +477,8 @@ NrRrcConnectionEstablishmentTestCase::CheckConnected(Ptr<NetDevice> ueDevice,
     uint16_t ueCellId = ueRrc->GetCellId();
     uint16_t ueDlBwp = ueRrc->GetPrimaryDlIndex();
     uint16_t ueUlBwp = ueRrc->GetPrimaryUlIndex();
-    uint8_t ueDlArfcn = ueNrDevice->GetArfcn(ueDlBwp);
-    uint8_t ueUlArfcn = ueNrDevice->GetArfcn(ueUlBwp);
+    uint32_t ueDlArfcn = ueNrDevice->GetArfcn(ueDlBwp);
+    uint32_t ueUlArfcn = ueNrDevice->GetArfcn(ueUlBwp);
     uint16_t ueImsi = ueNrDevice->GetImsi();
     uint8_t ueDlBandwidth = ueRrc->GetDlBandwidth();
     uint8_t ueUlBandwidth = ueRrc->GetUlBandwidth();
@@ -487,8 +487,8 @@ NrRrcConnectionEstablishmentTestCase::CheckConnected(Ptr<NetDevice> ueDevice,
     uint16_t gnbUlBwp = nrGnbDevice->GetArfcnBwpId(ueUlArfcn);
     uint8_t gnbDlBandwidth = nrGnbDevice->GetBwpDlBandwidth(gnbDlBwp);
     uint8_t gnbUlBandwidth = nrGnbDevice->GetBwpUlBandwidth(gnbUlBwp);
-    uint8_t gnbDlArfcn = nrGnbDevice->GetBwpArfcn(gnbDlBwp);
-    uint8_t gnbUlArfcn = nrGnbDevice->GetBwpArfcn(gnbUlBwp);
+    uint32_t gnbDlArfcn = nrGnbDevice->GetBwpArfcn(gnbDlBwp);
+    uint32_t gnbUlArfcn = nrGnbDevice->GetBwpArfcn(gnbUlBwp);
 
     NS_TEST_ASSERT_MSG_EQ(gnbRrc->HasCellId(ueCellId), true, "inconsistent CellId");
     NS_TEST_ASSERT_MSG_EQ(ueDlBandwidth, gnbDlBandwidth, "inconsistent DlBandwidth");
